Refactor app into focused views and helpers
