+++ src/lib/api.js
@@ -1,0 +1,29 @@
+export async function api(path, options = {}) {
+  const response = await fetch(path, {
+    credentials: 'include',
+    headers: {
+      'Content-Type': 'application/json',
+      ...(options.headers ?? {}),
+    },
+    ...options,
+  });
+
+  if (response.status === 204) {
+    return null;
+  }
+
+  const data = await response.json().catch(() => ({}));
+  if (!response.ok) {
+    throw new Error(data.error ?? `Request failed (${response.status})`);
+  }
+  return data;
+}
+
+export function reportClientError(message, stack = null, source = 'client') {
+  fetch('/api/logs/client-error', {
+    method: 'POST',
+    credentials: 'include',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify({ message, stack, source }),
+  }).catch(() => {});
+}
