+++ src/main.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App, { ErrorBoundary } from './App.jsx';
+import App from './App.jsx';
+import { ErrorBoundary } from './components/ErrorBoundary.jsx';
 import './styles.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
